build

+++ docs/public/assets/app.bundle.js
@@ -62641,9 +62641,9 @@ var Home = (_dec = (0, _reactReduxFirebase.firebaseConnect)([{ path: 'wallet', q
       data.createdAt = this.props.firebase.database.ServerValue.TIMESTAMP;
 
       if (data.type === 'income') {
-        this.props.firebase.update('/money', { total: parseInt(money.total) + parseInt(data.amount) });
+        this.props.firebase.update('/money', { total: parseFloat(money.total) + parseFloat(data.amount) });
       } else {
-        this.props.firebase.update('/money', { total: parseInt(money.total) - parseInt(data.amount) });
+        this.props.firebase.update('/money', { total: parseFloat(money.total) - parseFloat(data.amount) });
       }
       return this.props.firebase.push('/wallet', data);
       console.log('submitted', data);
@@ -62706,9 +62706,9 @@ var Home = (_dec = (0, _reactReduxFirebase.firebaseConnect)([{ path: 'wallet', q
       var money = this.props.money;
 
       if (data.type === 'income') {
-        this.props.firebase.update('/money', { total: parseInt(money.total) - parseInt(data.amount) });
+        this.props.firebase.update('/money', { total: parseFloat(money.total) - parseFloat(data.amount) });
       } else {
-        this.props.firebase.update('/money', { total: parseInt(money.total) + parseInt(data.amount) });
+        this.props.firebase.update('/money', { total: parseFloat(money.total) + parseFloat(data.amount) });
       }
       return this.props.firebase.remove('/wallet/' + id).catch(function (err) {
         console.error('Error removing wallet: ', err); // eslint-disable-line no-console
